// src/modules/disposal-points/enums/dp-entity-properties-db-names.enum.ts
/* DisposalPointEntityPropertiesDbNamesEnum */
export enum DisposalPointEntityPropertiesDbNamesEnum {
  ID = "dp_id",
  LATITUDE = "dp_latitude",
  LOCALITY_NAME = "dp_locality_name",
  LOCALITY_TYPE = "dp_locality_type",
  LONGITUDE = "dp_longitude",
  NEIGHBORHOOD = "dp_neighborhood",
  ACCEPTED_WASTE_CATEGORY = "dp_accepted_waste_category",
}